Create getAddressCoord.r
It takes a string as address, and send a query to Google Maps API. If the returned JSON is OK, the function returns a named numeric vector with 2 elements: "lat" and "lng"

It downloads the API's returned file to the hard drive temporarily, so it can be slow. However Google allows only 2500 queries/hour, so this is not the bottleneck.